and now with the right bins!!

+++ Colour_Dist_Test.py
@@ -33,7 +33,7 @@ def Update_DB_from_Color_Data(lowc, highc, mskewa, cur2):
 	print 'Number in bin: ', count
 	skewx=SkewG(x,1.8192627275,0.997793919871,-0.105487431764,0.117890808366)
 	print skewx, skewx/mskewa, x, count, int(count*skewx/mskewa)
-	setF=int(count*int(skewx/mskewa))
+	setF=int(count*skewx/mskewa)
 	print 'Updating Colour: ', x
 	#cur2.execute("UPDATE sn_mc SET colour_pass = False WHERE sn_id IN (SELECT sn_id from sn_mc where (color >=%s and color <%s) limit %s);",((float(lowc),float(highc),int(setF),)) )
 	cur2.execute("INSERT INTO colour_dis_sn_mc SELECT * from sn_mc where (color >=%s and color <%s) limit %s;",((float(lowc),float(highc),int(setF),)) )
